requirements.txt
boto==2.49.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Assets==0.12
flask-heroku==0.1.9
Flask-Scss==0.5
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.1.0
psycopg2==2.7.6.1
pyScss==1.3.5
python-docx==0.8.7
requests==2.20.1
six==1.11.0
SQLAlchemy==1.2.14
urllib3==1.24.1
webassets==0.12.1
Werkzeug==0.14.1
